Fix unit testing

+++ utilities/src/main/java/com/markp/dto/EmployeeDto.java
@@ -14,10 +14,11 @@ import java.util.List;
 @AllArgsConstructor
 public class EmployeeDto {
     private Long id;
+
     private String firstName;
     private String lastName;
     private String email;
-    private Long age;
+    private Integer age;
     private String address;
     private String contactNumber;
     private String employmentStatus;
